Add GlacierBed
This adds the GlacierBed class, an object for storing information about
the glacier bed. Initialisation can be done in two ways: Either with
single scalars for top, bottom and width or with altitude/width pairs
for greater control of the geometry. Also changes the glacier class to
accept the bed for initalisaiton.

Also contains some changes to the
plotting.