Reworked Edu MassBalance
The Edu MassBalance is now a full oggm.MassBalanceModel on its own. It
implement its own get_monthly_mb and get_annual_mb along with extensions
existing in the previous edu.MassBalance.

It is now possible to specify a mass balance with any number of
gradients by supplying the constructor with multiple gradients and
altitude breakpoints between them.

All changes are internal so no changes are needed for compatibility.