Glacier: copy of bed and mass balance.
Previously a glacier simply referenced the bed and glacier objects
passed at construction. This can cause some problems if multiple
glaciers are initialised from one set of bed and mass balance. When the
user then change an attribute of the mass balance through the glacier
object, this would cascade to all glacier based on that mass balance.
Now the glacier simply creates a copy of the bed and mass
balance, isolating them.