Passing partial expressions as values in change_attributes dict.
This makes is possible to pass partial mathematical expressions when
updating attributes through the change_attributes method of
GlacierCollection. This is done by passing a string e.g. '* 10' in the
value list of a key, This will multiply the current value by a factor of
10.

Partial expressions are evaluated in the new utility function expression_parser.

Adds test cases for both the expression_parser and the change_attributes
method.