Non-linear bed profiles.
This give the user more control over the slope of the glacier bed.
Either by supplying only a single slope or a sequence of slopes and
slope sections (specified by altitude intervals).

Behind the scenes it figures out how long (x-distance) each section has
to be to fulfill the slope, which is what the total glacier length and
the number of gridpoints is then based on. It then interpolates the
slope sections along the distance of the glacier.