First rules to _check_collection.
With the introduction of non-linear bed slopes, the collection need
logic to determine if glaciers can be put in the same collection. In
this initial case we check if the bed profiles are the same. This is to
keep plot methods working. Maybe this is something will do different in
the future. But for now I don't see a nice way to plot glaciers with
different slopes on top of each other.