Collection with different beds.
This changes how a collection deals with glaciers with different beds.
Before we didn't allow this. Now we simply raise an error in the plot
method. This way we can compare the history of glaciers with different
beds.

When a glacier, is added, we check if the beds are the same. If they are
not, we set the attribute collection_ok to false. This attribute can
then be used by a method to decide what to do.