Various quality improvements (#141)
* Change plot x-padding from fixed 2km to 2% of the glacier bed length.

* Logic update: Glacier grew out of domain

This change the behaviour when a glacier outgrows its domain.

- When progressing to a specified year, we now recurse back five year
  until the glacier fit the domain. This will return the glacier in its
  largest possible state.

- Progressing towards an equilibrium we now print a message saying that
  glacier ran out of space and then return. This is done instead of
  raising since it allows the collection to continue progressing even if
  one or more glaciers grow too large.

* Parallel collection progression

Enables the collection progression methods to progress the glaciers
in parallel. Makes it faster and a fun exercise.

* Add test for combining progress_to_year and progress_to_equilibrium.
History and state history previously did have the same length, (bug in
OGGM). Now fixed.

* Plot update. Change location of ELA annotation.

* Since OGGM now removes Nan years, we don't have to anymore.

* Temperature bias init check.

We now check that there isn't already a temperature bias scenario being
evaluated before the user can add a new one. Previous scenario has to be
completed before adding a new one.

* Reset method.

Adds a reset method to the glacier/surging_glacier and collection.
Mostly useful for collections. Resets the glaciers without the user
having to re-initialise them.

* Plot_state_history: Dealing with no state.

Adds a check to see if glacier have a state history. If not, we print
advise and return.

* Collection with different beds.

This changes how a collection deals with glaciers with different beds.
Before we didn't allow this. Now we simply raise an error in the plot
method. This way we can compare the history of glaciers with different
beds.

When a glacier, is added, we check if the beds are the same. If they are
not, we set the attribute collection_ok to false. This attribute can
then be used by a method to decide what to do.