Docs update: Api reference (#142)
* Convert some attributes and methods to private.

Some glacier attributes like current_state and model_state are never
interacted with by the user. The same is true for the method state().
To indicate this we make them "private". This also cleans up the
documentation.

* Docs update.

This adds a api reference of the new classes to the documentation. It
differs somewhat from the e.g. oggm.GlacierDir in that it also has a full
documentation of class methods. To achieve this an additional template
for classes is needed. Property setters are however not documented yet.

* Edits and new build structure

Co-authored-by: Fabien Maussion <fabien.maussion@uibk.ac.at>