Public current_state property.
This exposes the ``_current_state`` as a public property of ``Glacier``
and changes so that it is used where appropriate. It does not provide a
setter since this is only to be done internally through ``progress_to_year``
and ``progress_to_equilibrum``, which set the ``_current_state``
directly.

Also updates the documentation to include the new property.