temp_bias_series setter: Simplify setting custom mass balances.
It is much more simple to only pass an array of values to the setter.
The user no longer has to generate a correct dataframe with correct
years. This is now instead done in the setter. The user only has
to generate an array or list of biases.

Tests are updated.