Glens A, warnings and outputs (#152)
* Warnings for bad parameter choices.

This adds warnings when the user
- Sets a too low or high creep factor (Glen A)
- Sets a too low or high basal sliding.
- If the ELA is either above or below the altitude extent of the
  glacier.
- If spatial resolution is below 50m.

* Glens A temperature conversion.

This makes it possible to set the creep factor (Glens A) by supplying a
temperature below zero. The creep setter interprets the value and
converts it. If the value is already a creep factor (>0), it it set
directly.

* Accumulation area ratio

Adds the new attribute accumulation area ratio, as calculated in the
GlacierExplorer app. Also adds it to the summary table.

* Update tests for new bed angles. We can now set bed angles between -90
and 90 degrees.